+++ src/main/java/dandi/dandi/batch/weather/adapter/in/WeatherBatchController.java
@@ -1,5 +1,7 @@
-package dandi.dandi.batch.weather;
+package dandi.dandi.batch.weather.adapter.in;
 
+import dandi.dandi.batch.weather.application.service.WeatherBatchExecutor;
+import dandi.dandi.batch.weather.application.port.in.WeatherBatchRequest;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
